requirements.txt
requests==2.31.0
web3>=6.0.0
pandas==2.2.1
SQLAlchemy==2.0.28
psycopg2==2.9.9
python-dotenv==1.0.0
black==23.3.0
mypy==1.4.1
pylint==3.2.5
pytest==7.4.0
setuptools
pandas-stubs
types-psycopg2
types-requests
moralis
dune-client